+++ package.json
@@ -4,7 +4,7 @@
   "description": "Ambilight for iCue devices",
   "main": "main.js",
   "devDependencies": {
-    "electron": "^10.0.0",
+    "electron": "^13.6.2",
     "electron-builder": "^22.10.3",
     "electron-packager": "^15.2.0"
   },
